+++ apps/site/src/lib/blocks.ts
@@ -223,12 +223,7 @@ export const readBlocksFromDisk = async (): Promise<
 };
 
 // Blocks which are currently not compliant with the spec, and are thus misleading examples
-const blocksToHide = [
-  "@hash/callout",
-  "@hash/embed",
-  "@hash/header",
-  "@hash/paragraph",
-];
+const blocksToHide = ["@hash/embed"];
 
 /** Helps consistently hide certain blocks from the Hub and user profile pages */
 export const excludeHiddenBlocks = (