+++ apps/site/tests/integration/dashboard.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "playwright-test-coverage";
 
-import { resetDb } from "../shared/fixtures";
-import { login } from "../shared/nav";
+import { resetSite } from "../shared/fixtures.js";
+import { login } from "../shared/nav.js";
 
 test("dashboard page should not be accessible to guests", async ({ page }) => {
   await Promise.all([
@@ -13,7 +13,7 @@ test("dashboard page should not be accessible to guests", async ({ page }) => {
 });
 
 test("dashboard page should contain key elements", async ({ page }) => {
-  await resetDb();
+  await resetSite();
 
   await page.goto("/");
   await login({ page });