+++ apps/site/tests/integration/login-flow.test.ts
@@ -1,9 +1,9 @@
-import { Page } from "playwright";
+import type { Page } from "playwright";
 import { expect, test } from "playwright-test-coverage";
 
-import { readValueFromRecentDummyEmail } from "../shared/dummy-emails";
-import { resetDb } from "../shared/fixtures";
-import { login, openLoginModal, openMobileNav } from "../shared/nav";
+import { readValueFromRecentDummyEmail } from "../shared/dummy-emails.js";
+import { resetSite } from "../shared/fixtures.js";
+import { login, openLoginModal, openMobileNav } from "../shared/nav.js";
 
 const emailInputSelector = '[placeholder="claude\\@example\\.com"]';
 const loginButtonSelector = "button[type=submit]:has-text('Log In')";
@@ -29,7 +29,7 @@ const expectSignupButton = async ({
 };
 
 test.beforeEach(async () => {
-  await resetDb();
+  await resetSite();
 });
 
 test("login works for an existing user (via verification code)", async ({
@@ -80,7 +80,9 @@ test("login works for an existing user (via verification code)", async ({
 
   const logOutItem = accountDropdownPopover.locator(`span:has-text("Log Out")`);
   await expect(logOutItem).toBeVisible();
-  await expect(logOutItem).toHaveAttribute("href", "");
+  expect(await logOutItem.getAttribute("href")).toBeNull();
+  // @todo replace ↑ with ↓ when https://github.com/microsoft/playwright/issues/16270 is implemented
+  // await expect(logOutItem).not.hasAttribute("href");
 
   await logOutItem.click();