+++ packages/@blockprotocol/hook/src/hook-service-json.ts
@@ -1,0 +1,3 @@
+import hookServiceJson from "./hook-service.json" assert { type: "json" };
+
+export { hookServiceJson };