+++ packages/@blockprotocol/hook/src/react.ts
@@ -95,6 +95,7 @@ export const useHookEmbedderService = (
 
 type Hook<T extends HTMLElement> = {
   id: string | null;
+  cancel: () => void;
   teardown: (() => Promise<void>) | null;
   params: {
     service: HookBlockHandler | null;
@@ -188,32 +189,24 @@ export const useHook = <T extends HTMLElement>(
       return;
     }
 
-    const teardownPromise =
-      existingHookRef.current?.teardown?.().catch() ?? Promise.resolve();
+    const existingHookId = existingHookRef.current?.id;
+
+    existingHookRef.current?.cancel();
 
     if (node && service) {
       const controller = new AbortController();
 
-      /**
-       * Is this an update to the existing hook, or is it a whole new hook? The
-       * only param to the hook which can change without creating a new hook is
-       * the node. Any other change will result in a new hook being created
-       */
-      const reuseId =
-        existingHook &&
-        existingHook.service === service &&
-        existingHook.entityId === entityId &&
-        existingHook.path === path &&
-        existingHook.type === type;
-
       const hook: Hook<T> = {
-        id: reuseId ? existingHookRef.current?.id ?? null : null,
+        id: existingHookId ?? null,
         params: {
           service,
           type,
           entityId,
           path,
           node,
+        },
+        cancel() {
+          controller.abort();
         },
         async teardown() {
           if (controller.signal.aborted) {
@@ -253,45 +246,45 @@ export const useHook = <T extends HTMLElement>(
 
       existingHookRef.current = hook;
 
-      teardownPromise
-        .then(() => {
-          if (service.destroyed || controller.signal.aborted) {
-            return;
-          }
+      service
+        .hook({
+          data: {
+            hookId: hook.id,
+            entityId,
+            node,
+            type,
+            path,
+          },
+        })
+        .then((response) => {
+          if (!controller.signal.aborted) {
+            if (response.errors) {
+              const firstError = response.errors[0];
+              if (firstError?.code === "NOT_IMPLEMENTED") {
+                const teardown = fallbackRef.current(node);
 
-          return service
-            .hook({
-              data: {
-                hookId: hook.id,
-                entityId,
-                node,
-                type,
-                path,
-              },
-            })
-            .then((response) => {
-              if (!controller.signal.aborted) {
-                if (response.errors) {
-                  if (
-                    response.errors.length === 1 &&
-                    response.errors[0]?.code === "NOT_IMPLEMENTED"
-                  ) {
-                    const teardown = fallbackRef.current(node);
-
-                    hook.teardown = async () => {
-                      controller.abort();
-                      teardown?.();
-                    };
-                  } else {
-                    // eslint-disable-next-line no-console
-                    console.error(response.errors);
-                    throw new Error("Unknown error in hook");
-                  }
-                } else if (response.data) {
-                  hook.id = response.data.hookId;
+                hook.teardown = async () => {
+                  controller.abort();
+                  teardown?.();
+                };
+              } else if (firstError?.code === "NOT_FOUND") {
+                const errMsg = `Hook with id ${hook.id} was not found by embedding application`;
+                if (node === null) {
+                  // don't throw if the request was for hook deletion – the embedding app can't find the hook, things can continue
+                  // eslint-disable-next-line no-console -- useful for debugging
+                  console.warn(`${errMsg} – no hook to remove`);
+                } else {
+                  throw new Error(errMsg);
                 }
+              } else {
+                // eslint-disable-next-line no-console
+                console.error(response.errors);
+                throw new Error("Unknown error in hook");
               }
-            });
+            } else if (response.data) {
+              hook.id = response.data.hookId;
+            }
+          }
         })
         .catch((err) => {
           catchError(() => {