+++ packages/@blockprotocol/type-system/test/property-type.test.ts
@@ -5,7 +5,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/age/v/1",
     title: "Age",
-    pluralTitle: "Ages",
     oneOf: [
       {
         $ref: "https://blockprotocol.org/@blockprotocol/types/data-type/number/v/1",
@@ -16,7 +15,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/contact-information/v/1",
     title: "Contact Information",
-    pluralTitle: "Contact Information",
     oneOf: [
       {
         type: "object",
@@ -40,7 +38,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/contrived-property/v/1",
     title: "Contrived Property",
-    pluralTitle: "Contrived Properties",
     oneOf: [
       {
         $ref: "https://blockprotocol.org/@blockprotocol/types/data-type/number/v/1",
@@ -62,7 +59,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/favorite-quote/v/1",
     title: "Favorite Quote",
-    pluralTitle: "Favorite Quotes",
     oneOf: [
       {
         $ref: "https://blockprotocol.org/@blockprotocol/types/data-type/text/v/1",
@@ -73,7 +69,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/interests/v/1",
     title: "Interests",
-    pluralTitle: "Interests",
     oneOf: [
       {
         type: "object",
@@ -101,7 +96,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/name/v/1",
     title: "Name",
-    pluralTitle: "Names",
     oneOf: [
       {
         $ref: "https://blockprotocol.org/@blockprotocol/types/data-type/text/v/1",
@@ -112,7 +106,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/numbers/v/1",
     title: "Numbers",
-    pluralTitle: "Numbers",
     oneOf: [
       {
         type: "array",
@@ -130,7 +123,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/text/v/1",
     title: "Text",
-    pluralTitle: "Text",
     oneOf: [
       {
         $ref: "https://blockprotocol.org/@blockprotocol/types/data-type/text/v/1",
@@ -141,7 +133,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/user-id/v/1",
     title: "User ID",
-    pluralTitle: "User IDs",
     oneOf: [
       {
         $ref: "https://blockprotocol.org/@blockprotocol/types/data-type/text/v/1",
@@ -152,7 +143,6 @@ const propertyTypes: PropertyType[] = [
     kind: "propertyType",
     $id: "https://blockprotocol.org/@alice/types/property-type/user-id/v/2",
     title: "User ID",
-    pluralTitle: "User IDs",
     oneOf: [
       {
         $ref: "https://blockprotocol.org/@blockprotocol/types/data-type/text/v/1",
@@ -168,31 +158,11 @@ const propertyTypes: PropertyType[] = [
 // validation method
 const invalidPropertyTypes: [string, PropertyType, ParsePropertyTypeError][] = [
   [
-    "empty one of",
-    {
-      kind: "propertyType",
-      $id: "https://blockprotocol.org/@blockprotocol/types/property-type/broken/v/1",
-      title: "Broken",
-      pluralTitle: "Broken",
-      oneOf: [],
-    },
-    {
-      reason: "InvalidOneOf",
-      inner: {
-        reason: "ValidationError",
-        inner: {
-          type: "EmptyOneOf",
-        },
-      },
-    },
-  ],
-  [
     "non-integer version",
     {
       kind: "propertyType",
       $id: "https://blockprotocol.org/@blockprotocol/types/property-type/broken/v/1.4",
       title: "Broken",
-      pluralTitle: "Broken",
       oneOf: [
         {
           $ref: "https://blockprotocol.org/@blockprotocol/types/data-type/number/v/1",
@@ -212,7 +182,6 @@ const invalidPropertyTypes: [string, PropertyType, ParsePropertyTypeError][] = [
       kind: "propertyType",
       $id: "https://  /broken/v/1",
       title: "Broken",
-      pluralTitle: "Broken",
       oneOf: [
         {
           $ref: "https://blockprotocol.org/@blockprotocol/types/data-type/number/v/1",
@@ -236,7 +205,6 @@ const invalidPropertyTypes: [string, PropertyType, ParsePropertyTypeError][] = [
       kind: "propertyType",
       $id: "https://blockprotocol.org/@blockprotocol/types/property-type/broken/v/1",
       title: "Broken",
-      pluralTitle: "Broken",
       oneOf: [
         {
           $ref: "im a broken ref haha /v/1",
@@ -266,7 +234,6 @@ const invalidPropertyTypes: [string, PropertyType, ParsePropertyTypeError][] = [
       kind: "propertyType",
       $id: "https://blockprotocol.org/@blockprotocol/types/property-type/broken/v/1",
       title: "Broken",
-      pluralTitle: "Broken",
       oneOf: [
         {
           type: "object",
@@ -317,12 +284,18 @@ const brokenTypes: [any, ParsePropertyTypeError][] = [
   [
     {
       kind: "propertyType",
-      $id: "https://blockprotocol.org/@blockprotocol/types/property-type/age/v/1",
-      title: "Age",
+      $id: "https://blockprotocol.org/@blockprotocol/types/property-type/broken/v/1",
+      title: "Broken",
+      oneOf: [],
     },
     {
-      reason: "InvalidJson",
-      inner: "missing field `pluralTitle` at line 1 column 114",
+      reason: "InvalidOneOf",
+      inner: {
+        reason: "ValidationError",
+        inner: {
+          type: "EmptyOneOf",
+        },
+      },
     },
   ],
 ];