+++ packages/@blockprotocol/hook/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @blockprotocol/hook
 
+## 0.0.8
+
+### Patch Changes
+
+- [#707](https://github.com/blockprotocol/blockprotocol/pull/707) [`62396fa`](https://github.com/blockprotocol/blockprotocol/commit/62396fa3cc77bb1e70c5501079fe0f9b2cf4ef23) Thanks [@CiaranMn](https://github.com/CiaranMn)! - handle hook NOT_FOUND error. don't tear down when updating hook
+
+## 0.0.7
+
+### Patch Changes
+
+- [#676](https://github.com/blockprotocol/blockprotocol/pull/676) [`97d8751`](https://github.com/blockprotocol/blockprotocol/commit/97d8751a8e700293869d86c6062db9e88c1e24af) Thanks [@CiaranMn](https://github.com/CiaranMn)! - export hook-service.json. add error codes. bug fix in useHook
+
+## 0.0.6
+
+### Patch Changes
+
+- [#613](https://github.com/blockprotocol/blockprotocol/pull/613) [`0afbd33`](https://github.com/blockprotocol/blockprotocol/commit/0afbd330a674765b5c9c1c0c19c1acb8be6a45e7) Thanks [@kachkaev](https://github.com/kachkaev)! - Fix typo in README
+
+- [#647](https://github.com/blockprotocol/blockprotocol/pull/647) [`14ab891`](https://github.com/blockprotocol/blockprotocol/commit/14ab891bc1cdd10fce2b1e8651fa897323c64e15) Thanks [@CiaranMn](https://github.com/CiaranMn)! - add entityId to @blockprotocol/hook
+
 ## 0.0.5
 
 ### Patch Changes